# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patient Medicine Tracker (PMM) API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://pmm.oloma.dev/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh the token",
                "operationId": "auth_refresh",
                "requestBody": {
                    "description": "Token refresh request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserObject"
                                        },
                                        "avatar": {
                                            "$ref": "#/components/schemas/AvatarObject"
                                        },
                                        "expiresAt": {
                                            "description": "Expiration date of token",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Response: token is expired"
                    }
                }
            }
        },
...
        "/authorization/roles/findAllByPaging": {
            "get": {
                "tags": [
                    "Authorization Roles"
                ],
                "summary": "Find all roles by pagination",
                "operationId": "authorizationRoles_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesFindAllByPaging"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/authorization/roles/findAll": {
            "get": {
                "tags": [
                    "Authorization Roles"
                ],
                "summary": "Find all roles",
                "operationId": "authorizationRoles_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/authorization/roles/findOneById/{roleId}": {
            "get": {
                "tags": [
                    "Authorization Roles"
                ],
                "summary": "Find item data",
                "operationId": "authorizationRoles_findOneById",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesFindOneById"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/modules/create": {
            "post": {
                "tags": [
                    "Modules"
                ],
                "summary": "Create a new module",
                "operationId": "modules_create",
                "requestBody": {
                    "description": "Create a new role",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuleSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/modules/delete/{permId}": {
            "delete": {
                "tags": [
                    "Modules"
                ],
                "summary": "Delete module",
                "operationId": "modules_delete",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "Module uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RolesFindAllByPaging": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RolesFindAllByPagingObject"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "RolesFindAllByPagingObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RolesFindOneById": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/RolesFindOneByIdObject"
                    }
                },
                "type": "object"
            },
            "RolesFindOneByIdObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string"
                    },
                    "rolePermissions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "method": {
                                    "type": "string"
                                },
                                "route": {
                                    "type": "string"
                                },
                                "action": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "roleUsers": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "firstname": {
                                    "type": "string"
                                },
                                "lastname": {
                                    "type": "string"
                                },
                                "isActive": {
                                    "type": "number"
                                },
                                "createdAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "AvatarObject": {
                "properties": {
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserObject": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ModuleSave": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}